Correct a comment.
I had wrongly believed my TYPECHECK macro double-evaluated one of its
arguments and hence would cause side effects to happen twice. But in
fact I've just realised that although it double-_expands_ the
argument, it doesn't double-_evaluate_ it: the two expansions occur in
mutually exclusive branches of a ?:, and hence cannot both be
executed.

So I've removed the comment that says my macro is rubbish. My macro is
in fact great :-)

[originally from svn r9328]
[this svn revision also touched charset,filter,timber]