New build script for testing purposes.
'bob -s Buildscr.test' builds Halibut, runs it over inputs/test.but to
produce all the supported output formats, and delivers them all into
the output directory. Including Windows HTML Help, for which it has to
do a special run of the HTML back end with extra options and then get
a Windows box to run hhc (hence this having to be a Buildscr).