Add missing initialisations in the 'word' structure.
The 'breaks' and 'aux' fields were filled in rather inconsistently at
various places where a word is created - especially the outlying ones
that manufacture pieces of document during internal processing of
contents, index, bibliography, cross-references etc rather than
directly from the input file. This has never led to any user-visible
behaviour change that I've noticed, but it made a lot of annoying
noise in the valgrind output, which got in my way last week when I was
trying to debug the CHM generation.