Ignore leading whitespace in an input file.
A user points out (Debian bug #898923) that Halibut copes fine if an
input file starts with two or more blank lines, or zero, followed by a
paragraph-initial keyword such as \cfg, but gives a parse error if it
starts with exactly one. The same failure happens if the file starts
with a space and then \cfg.

That's because a sequence of leading whitespace containing at least
two \n is returned as a tok_eop (with an empty paragraph preceding it
which the parser discards), and no whitespace characters at all don't
lex as any token, but if there's a sequence of leading whitespace that
is non-empty and contains fewer than two \n, then it lexes as
tok_white which the parser wasn't expecting. This only causes trouble
in the first paragraph of the file, because after that, leading
whitespace in any subsequent paragraph is folded into the previous
tok_eop.