Fix misplaced parens in bk_whlp.c.
To compare a character case-insensitively with 'h', you have to write
tolower(c) != 'h', not tolower(c != 'h'). Ahem. That would have caused
an uppercase .HLP extension in the configuration not to be recognised
as valid, and Halibut would have put an extra .hlp on the end just in
case.