Revise tree234 comparison API.
Comparison functions now take a third context parameter, in case they
need to depend on something not accessible as a global variable. As
yet this functionality is unused in Halibut: all context parameters so
far used are NULL.

This means that newtree234, find234 and split234, which always took a
comparison function argument, now have to take a context pointer as
well. That's particularly awkward for the find and split functions,
which I think are likely to _often_ be called with NULL, NULL to use
the tree's built-in version; so I've also revised the API of those
functions to omit the comparison function, and made the versions that
do take a special compare function be 'findcmp234' and friends.

While I'm at it, the void * element pointers are now const void *,
which is a long-overdue cleanup.