Refactor global variables into a 'psdata' struct.
No functional change, but this restores an invariant that the Halibut
code base originally had, which is that there are no mutable global
variables - everything that modifies state is passed that state as a
parameter, and ultimately, everything is either allocated on the heap
or some function's stack.

(There was never an especially good reason for this property. I had
the vague idea at one time that it might be useful to have the Halibut
code callable as sone kind of library, but never really came up with a
good use for that in turn. However, now that the code _is_ mostly set
up that way, I don't want to gratuitously lose it.)

Ben's work on PS font management years ago broke that invariant,
probably because I completely forgot to mention it to him. Now
restored, by moving his cache of PS glyph names and global list of
loaded fonts into a 'psdata' structure which is passed in to
everything that needs to access it.

It would be nice to be able to treat 'psdata' as a detail local to the
paper backend cluster - have bk_paper put it in the 'document'
structure, and bk_ps and bk_pdf read it out again. But it doesn't work
that way, because PS fonts are loaded during the input phase, so the
psdata has to already exist and be populated before bk_paper even
runs. So psdata is a top-level thing that main() has to know about.