Remove the libcharset submodule.
I've gradually become convinced by the arguments that say git subtree
is preferable to git submodule. I went with submodules when I
originally converted Halibut and libcharset from svn, but there's no
reason that has to be a lifetime commitment: we can just get rid of
the submodule, and then re-import the same commit of the same
repository as a subtree.

This commit breaks the build, by removing the charset subdirectory.
The next commit will make it work again.