gcc warning fix: avoid strncpy in whlp_make_fontsection.
Annoyingly, gcc 10's -Wstringop-truncation has made strncpy
essentially useless.

The one thing it's actually _useful_ for is filling up fixed-length
string fields with the semantics 'may fill the whole buffer if the
string is maximally long, but must be NUL-terminated if it's any
shorter', such as the ones in Unix structures like 'struct passwd' or
'struct dirent', or (in this case) one of the fields in a Windows .HLP
file subtable. In these situations, strncpy(buf, string, bufsize)
fills in the buffer just as you'd like, with the bonus of avoiding
leaving nondeterministic cruft after the NUL if the string is shorter
than n-1 bytes.

But it's in exactly that situation where gcc warns about it, because
it might not NUL-terminate the output buffer. gcc would rather you
used it in situations where you always had to NUL-terminate your
string, in which case it's not what you wanted anyway because the
trailing NUL-padding makes it too slow!