Fix null dereference on an empty index term.
When read_file processes the case-insensitive index directive \ii, it
calls ustrlow() on indexstr.text (the rdstring where the text inside
that directive has been accumulated), and then appends a zero byte to
pass it to index_merge.

But if you write a completely empty \ii{}, then indexstr.text was NULL
at the point where ustrlow() tried to dereference it. That's not a
very useful thing to do, but we can at least not segfault. Ensure
we've allocated a string by that point, even if it was empty.