Allow building the docs even in a cross-compile.
Colin Watson points out that in the context of a Debian package build,
cross-compiling doesn't just need to generate some binaries: your
cross-build wants to go all the way to 'make install' so that it can
set up the entire file structure for the cross-built package. So, if
the cross-build refuses to build the docs, then the docs don't get
installed, and the Debian machinery has to do some nasty bodging.

Previously this worked OK because the docs Makefile was separate and
took a HALIBUT parameter to override which binary was used for
building the docs. So you could just add a parameter to your 'make -C
doc' to point it at the Halibut binary from the native build you were
also doing.

Now you can do 'cmake -DNATIVE_HALIBUT=/some/path/to/halibut' in a
cross-build, and then you get the docs built using _that_ Halibut, so
the cross-build ends up with a full set of files and can 'make install'.